+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to the public API will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Changes to be tagged
+### Added
+- charaGraphEx to servant assets.
+- skillAdd to skill.
 
 ## 5.69.0 - 2021-08-10
 ### Added